+++ app/build.gradle
@@ -81,9 +81,6 @@ dependencies {
     // Coil
     implementation 'io.coil-kt:coil-compose:2.4.0'
 
-    // Moshi
-    implementation 'com.squareup.moshi:moshi-kotlin:1.13.0'
-    implementation 'com.squareup.retrofit2:converter-moshi:2.9.0'
 
     // OkHttp
     implementation("com.squareup.okhttp3:okhttp:5.0.0-alpha.11")
