+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/data/remote/OpenFoodApi.kt
@@ -4,13 +4,14 @@ import ar.edu.unlam.mobile.scaffold.data.remote.dto.SearchDto
 import retrofit2.http.GET
 import retrofit2.http.Query
 
+
 interface OpenFoodApi {
 
     @GET("cgi/search.pl?search_simple=1&json=1&action=process&fields=product_name,nutriments,image_front_thumb_url")
     suspend fun searchFood(
         @Query("search_terms") query: String,
         @Query("page") page: Int,
-    @Query("page_size") pageSize: Int,
+        @Query("page_size") pageSize: Int,
     ): SearchDto
 
     companion object {
