+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/data/remote/dto/Nutriments.kt
@@ -1,15 +1,14 @@
 package ar.edu.unlam.mobile.scaffold.data.remote.dto
 
-import com.squareup.moshi.Json
+import com.google.gson.annotations.SerializedName
 
 data class Nutriments(
-    @field:Json(name = "carbohydrates_100g")
+    @SerializedName("carbohydrates_100g")
     val carbohydrates100g: Double,
-    @field:Json(name = "energy-kcal_100g")
+    @SerializedName("energy-kcal_100g")
     val energyKcal100g: Double,
-    @field:Json(name = "fat_100g")
+    @SerializedName("fat_100g")
     val fat100g: Double,
-    @field:Json(name = "proteins_100g")
+    @SerializedName("proteins_100g")
     val proteins100g: Double,
-
 )
