+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/data/remote/dto/SearchDto.kt
@@ -1,5 +1,6 @@
 package ar.edu.unlam.mobile.scaffold.data.remote.dto
 
+
 data class SearchDto(
     val products: List<Product>,
 )
