--- app/src/main/java/ar/edu/unlam/mobile/scaffold/ui/Color.kt
@@ -1,14 +1,0 @@
-package ar.edu.unlam.mobile.scaffold.ui
-
-import androidx.compose.ui.graphics.Color
-
-val BrightGreen = Color(0xFF00C713)
-val DarkGreen = Color(0xFF00790C)
-val Orange = Color(0xFFFFAA00)
-val CarbColor = Color(0xFFEEFF00)
-val ProteinColor = Orange
-val FatColor = Color(0xFFF44336)
-val LightGray = Color(0xFF808080)
-val MediumGray = Color(0xFF404040)
-val DarkGray = Color(0xFF202020)
-val TextWhite = Color(0xFFEEEEEE)
