+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/ui/components/DaySelector.kt
@@ -1,6 +1,7 @@
 package ar.edu.unlam.mobile.scaffold.ui.components
 
 import android.os.Build
+import android.util.Log
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
@@ -26,6 +27,8 @@ fun DaySelector(
     onNextDayClick: () -> Unit,
     modifier: Modifier = Modifier,
 ) {
+    // TODO borrar esto
+    Log.i("mytag","Se lanzo un DaySelector")
     Row(
         modifier = modifier,
         horizontalArrangement = Arrangement.SpaceBetween,
