+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/ui/screens/SearchScreen.kt
@@ -1,7 +1,6 @@
 package ar.edu.unlam.mobile.scaffold.ui.screens
 
 import android.os.Build
-import android.util.Log
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -29,7 +28,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import ar.edu.unlam.mobile.scaffold.R
 import ar.edu.unlam.mobile.scaffold.core.util.UiEvent
 import ar.edu.unlam.mobile.scaffold.domain.model.MealType
-import ar.edu.unlam.mobile.scaffold.ui.LocalSpacing
+import ar.edu.unlam.mobile.scaffold.ui.theme.LocalSpacing
 import ar.edu.unlam.mobile.scaffold.ui.components.SearchTextField
 import ar.edu.unlam.mobile.scaffold.ui.components.TrackableFoodItem
 import java.time.LocalDate
@@ -48,6 +47,7 @@ fun SearchScreen(
 ) {
     val spacing = LocalSpacing.current
     val state = viewModel.state
+
     val context = LocalContext.current
     val keyboardController = LocalSoftwareKeyboardController.current
     LaunchedEffect(key1 = keyboardController) {
@@ -83,7 +83,6 @@ fun SearchScreen(
             onSearch = {
                 keyboardController?.hide()
                 viewModel.onEvent(SearchEvent.OnSearch)
-                Log.i(null,"Metodo search llamado")
             },
             onFocusChange = {
                 viewModel.onEvent(SearchEvent.OnSearchFocusChange(it.isFocused))
