+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/ui/screens/TrackerOverviewViewmodel.kt
@@ -74,6 +74,9 @@ class TrackerOverviewViewmodel @Inject constructor(
         }
     }
 
+    /**
+     * Devuelve las comidas seguidas por fecha
+     */
     private fun refreshFoods() {
         getFoodsForDate?.cancel()
         getFoodsForDate = trackerUseCases
