+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/ui/screens/activity/ActivityScreen.kt
@@ -22,10 +22,13 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import ar.edu.unlam.mobile.scaffold.R
 import ar.edu.unlam.mobile.scaffold.core.util.UiEvent
 import ar.edu.unlam.mobile.scaffold.domain.model.ActivityLevel
-import ar.edu.unlam.mobile.scaffold.ui.LocalSpacing
+import ar.edu.unlam.mobile.scaffold.ui.theme.LocalSpacing
 import ar.edu.unlam.mobile.scaffold.ui.components.ActionButton
 import ar.edu.unlam.mobile.scaffold.ui.components.SelectableButton
 
+/**
+ * Pantalla para establecer el nivel de la actividad fisica de la persona.
+ */
 @Composable
 fun ActivityScreen(
     onNextClick: () -> Unit,
