+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/ui/theme/Color.kt
@@ -2,10 +2,13 @@ package ar.edu.unlam.mobile.scaffold.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
+val BrightGreen = Color(0xFF00C713)
+val DarkGreen = Color(0xFF00790C)
+val Orange = Color(0xFFFFAA00)
+val CarbColor = Color(0xFFEEFF00)
+val ProteinColor = Orange
+val FatColor = Color(0xFFF44336)
+val LightGray = Color(0xFF808080)
+val MediumGray = Color(0xFF404040)
+val DarkGray = Color(0xFF202020)
+val TextWhite = Color(0xFFEEEEEE)
